+++ PPE/backend/.env
@@ -1,1 +1,2 @@
 PORT = 3000
+VITE_SERVER_URL=http://localhost:3000
