+++ PPE/frontend/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.4.0",
     "recharts": "^2.15.1",
     "socket.io-client": "^4.8.1"
   },
